Replace custom caching logic with external rust caching action
This removes the complex logic behind the finer grained cache layers and
replaces it with a pre-built caching action from the marketplace.

This leads to slightly more time spent compiling but less time spent
restoring caches resulting in an equal CI runtime (but much less complexity).

CI runs for comparison:
* old: https://github.com/dfinity/internet-identity/actions/runs/5065481740/jobs/9094102129
* new: https://github.com/dfinity/internet-identity/actions/runs/5065432317/jobs/9094005126