Opt out /metrics endpoint from certification
This PR adds a certified opt-out for the /metrics endpoint
so that it can be accessed using the non-raw domain.

Note: this does not work (yet) in practice, as this feature has
not been implemented in the SW. However, the validation library
already allows checking that the opt-out is setup correctly.